requirements.txt
ruamel.yaml
jsonschema
deepmerge